cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kurultai"
version = "0.1.0"
description = "Kurultai CLI - AI Skill Registry and Management Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Kurultai Team"}
]
keywords = ["cli", "ai", "skills", "registry", "management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "GitPython>=3.1",
    "pyyaml>=6.0",
    "semantic-version>=2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
kurultai = "kurultai.cli:main"

[project.urls]
Homepage = "https://github.com/kurultai/kurultai-cli"
Repository = "https://github.com/kurultai/kurultai-cli"
Issues = "https://github.com/kurultai/kurultai-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=kurultai --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/kurultai"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
